Changed apt-get to apt
Updated clean-up script to use apt instead of apt-get.
apt is the newer tool and it works better than simple apt-get. If apt-get is used, not all security patches get installed and img_check.sh will fail. Note that apt-get can also be used, but is has to be enhanced with additional flags like  --with-new-pkgs and etc. I think apt is the more appropriate solution.