Improve DO agent check
The DO agent check, checks for the presence of a directory under
/opt/digitalocean. The name of this directly slightly varies between
various distributions. Rather than check a sub-directory, simply check
for the existence of the /opt/digitalocean directory which is consistent
between distributions and does not exist if the DO agent or monitoring
are installed. Also, update the instructions for properly cleaning up if
removing a previously installed agent and include Debian.